Trying to add diff algorithm to server but fail
I have decided that it would be way easier to write a web page instead
of fix the eclipse plugin.

After this commit, i will make a "writer.htm" instead of working further
on. Thus i will only be working in javascript & python. The writer
client will have the ability to save to anywhere on the pc via the
server. The server will always create a backup copy of the file to work
with then commit changes to the original when I am happy that it is
stable.