Basic functionality up and going
Fixed up tabs and added "close" and "refresh"
Working on "replace" functionality.. Currently it is fucking up because
ACE is a retard and "onchange" is triggered when i do editor.setValue()
FUCKING RETARDED. This has created all kinds of problems. ACE is really
not well thought out once you get into the grit of it.